Cache results for duration of override
Purging of cache on record change was introduced in 147b945 so it is
safe for us to cache a throttle override until its current expiration
time on successful lookup. The cache time for a negative lookup is kept
at one hour, not because a longer duration would be unsafe, but because
the key space is incredibly large (2^32 + 2^128).

Change-Id: I049d20f93b0d66f4109778e4e664d2f73c88ae24